[fix] Remove race condition in met_forcing
Previously, the device variables windX_d, windY_d are used in the kernel
loop directive but seem to be treated as shared variables, causing race
conditions. We now create a separate kernel for this loop to use local
variables.

Note that the "device" attribute was needed in the former because the
ApplyDirectionalWindReduction device subroutine modified those
variables.