Fix internal_edge_hydro bug in GPU code (#38)
In the GPU code, the `dry_flux` branch seems to have been optimized
away when using `-O1` or above. This causes significant deviations to
CPU code when running on meshes with wetting and drying. To avoid this,
the numerical flux function now returns a vector argument and is now a
pure function. Gravity modification has also been added back.

The default relative error for the test cases is now 1e-7 (translates to
7 digits matching).